Add support to use keys and certificates from PKCS#11 provider
In OpenSSL < 3.0, the modularity was provided by mechanism
called "engines". This is supported in curl, but the engines
got deprecated with OpenSSL 3.0 in favor of more versatile
and (from use point of view) complicated providers.

This adds a support for OpenSSL Providers, to use PKCS#11
keys, namely through the pkcs11 provider. This is done using
similar approach as the engines and this is automatically built
in when the OpenSSL 3 and newer is used.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>